+++ src/pages/LoginView/LoginView.jsx
@@ -1,5 +1,5 @@
 import SignIn from "./LoginForm";
-import styles from "./LoginScreen.module.css";
+import styles from "./LoginView.module.css";
 
 const LoginScreen = () => {
   return (
